Fix printing integer type images (bug #51558)
*gl2ps-print.cc(gl2ps_renderer::draw_pixels): get the correct maximum value for
 uint8_t and uint16_t types.

+++ libinterp/corefcn/gl2ps-print.cc
@@ -763,7 +763,7 @@ namespace octave
 
     OCTAVE_LOCAL_BUFFER (float, tmp_data, 3*w*h);
 
-    static const float maxval = std::numeric_limits<float>::max ();
+    static const float maxval = std::numeric_limits<uint8_t>::max ();
 
     for (int i = 0; i < 3*w*h; i++)
       tmp_data[i] = data[i] / maxval;
@@ -778,7 +778,7 @@ namespace octave
 
     OCTAVE_LOCAL_BUFFER (float, tmp_data, 3*w*h);
 
-    static const float maxval = std::numeric_limits<float>::max ();
+    static const float maxval = std::numeric_limits<uint16_t>::max ();
 
     for (int i = 0; i < 3*w*h; i++)
       tmp_data[i] = data[i] / maxval;
